chore: rename `tools` to `tool`
