+++ lua/modules/configs/tool/trouble.lua
